+++ src/constants/theme.ts
@@ -291,5 +291,5 @@ export const Spacing = {
   six: 64,
 } as const;
 
-export const BottomTabInset = Platform.select({ ios: 50, android: 80 }) ?? 0;
+export const BottomTabInset = Platform.select({ ios: 78, android: 92 }) ?? 0;
 export const MaxContentWidth = 800;